refactor(commands): unify command group registration pattern
Introduce consistent `register_*` pattern for command aggregation:

**Flow commands**:
- Extract management commands (update, bind, list-deleted, restore) to flow_manage.py
- Add register_manage_commands() to flow_manage.py
- Add register_lifecycle_commands() to flow_lifecycle.py
- Add register_status_commands() to flow_status.py
- Refactor flow.py to call register_* functions

**Handoff commands**:
- Add register_read_commands() to handoff_read.py
- Add register_write_commands() to handoff_write.py
- Refactor handoff.py to call register_* functions

**Tests**:
- Update imports to patch correct module paths (flow → flow_manage)

This unifies the registration pattern across all command groups,
improving maintainability without changing CLI behavior.

Closes #348
