+++ tests/vibe3/commands/test_task_management_bind.py
@@ -11,13 +11,15 @@ runner = CliRunner(env={"NO_COLOR": "1"})
 
 def test_flow_bind_supports_related_role() -> None:
     """Test flow bind successfully binds a task to the current flow."""
-    with patch("vibe3.commands.flow.TaskService", create=True) as task_service_cls:
+    with patch(
+        "vibe3.commands.flow_manage.TaskService", create=True
+    ) as task_service_cls:
         task_service = MagicMock()
         task_service_cls.return_value = task_service
 
         flow_service = MagicMock()
         flow_service.get_current_branch.return_value = "task/demo"
-        with patch("vibe3.commands.flow.FlowService", return_value=flow_service):
+        with patch("vibe3.commands.flow_manage.FlowService", return_value=flow_service):
             result = runner.invoke(flow_app, ["bind", "219"])
 
     assert result.exit_code == 0
@@ -28,13 +30,15 @@ def test_flow_bind_supports_related_role() -> None:
 
 def test_flow_bind_with_task_role() -> None:
     """flow bind 220 --role task should bind as task."""
-    with patch("vibe3.commands.flow.TaskService", create=True) as task_service_cls:
+    with patch(
+        "vibe3.commands.flow_manage.TaskService", create=True
+    ) as task_service_cls:
         task_service = MagicMock()
         task_service_cls.return_value = task_service
 
         flow_service = MagicMock()
         flow_service.get_current_branch.return_value = "task/demo"
-        with patch("vibe3.commands.flow.FlowService", return_value=flow_service):
+        with patch("vibe3.commands.flow_manage.FlowService", return_value=flow_service):
             result = runner.invoke(flow_app, ["bind", "220", "--role", "task"])
 
     assert result.exit_code == 0
@@ -45,13 +49,15 @@ def test_flow_bind_with_task_role() -> None:
 
 def test_flow_bind_with_related_role() -> None:
     """flow bind 219 --role related should bind as related."""
-    with patch("vibe3.commands.flow.TaskService", create=True) as task_service_cls:
+    with patch(
+        "vibe3.commands.flow_manage.TaskService", create=True
+    ) as task_service_cls:
         task_service = MagicMock()
         task_service_cls.return_value = task_service
 
         flow_service = MagicMock()
         flow_service.get_current_branch.return_value = "task/demo"
-        with patch("vibe3.commands.flow.FlowService", return_value=flow_service):
+        with patch("vibe3.commands.flow_manage.FlowService", return_value=flow_service):
             result = runner.invoke(flow_app, ["bind", "219", "--role", "related"])
 
     assert result.exit_code == 0
@@ -62,13 +68,15 @@ def test_flow_bind_with_related_role() -> None:
 
 def test_flow_bind_with_dependency_role() -> None:
     """flow bind 218 --role dependency should bind as dependency."""
-    with patch("vibe3.commands.flow.TaskService", create=True) as task_service_cls:
+    with patch(
+        "vibe3.commands.flow_manage.TaskService", create=True
+    ) as task_service_cls:
         task_service = MagicMock()
         task_service_cls.return_value = task_service
 
         flow_service = MagicMock()
         flow_service.get_current_branch.return_value = "task/demo"
-        with patch("vibe3.commands.flow.FlowService", return_value=flow_service):
+        with patch("vibe3.commands.flow_manage.FlowService", return_value=flow_service):
             result = runner.invoke(flow_app, ["bind", "218", "--role", "dependency"])
 
     assert result.exit_code == 0
@@ -79,7 +87,9 @@ def test_flow_bind_with_dependency_role() -> None:
 
 def test_flow_bind_with_explicit_branch_option() -> None:
     """flow bind --branch task/other 219 should bind to the specified branch."""
-    with patch("vibe3.commands.flow.TaskService", create=True) as task_service_cls:
+    with patch(
+        "vibe3.commands.flow_manage.TaskService", create=True
+    ) as task_service_cls:
         task_service = MagicMock()
         task_service_cls.return_value = task_service
 
@@ -87,7 +97,7 @@ def test_flow_bind_with_explicit_branch_option() -> None:
         flow_service.get_current_branch.return_value = "task/demo"
         flow_service.get_flow_status.return_value = MagicMock()
         flow_service._is_main_branch.return_value = False
-        with patch("vibe3.commands.flow.FlowService", return_value=flow_service):
+        with patch("vibe3.commands.flow_manage.FlowService", return_value=flow_service):
             result = runner.invoke(flow_app, ["bind", "--branch", "task/other", "219"])
 
     assert result.exit_code == 0
@@ -98,13 +108,15 @@ def test_flow_bind_with_explicit_branch_option() -> None:
 
 def test_flow_bind_with_explicit_protected_branch_fails() -> None:
     """flow bind --branch main should fail before linking issues."""
-    with patch("vibe3.commands.flow.TaskService", create=True) as task_service_cls:
+    with patch(
+        "vibe3.commands.flow_manage.TaskService", create=True
+    ) as task_service_cls:
         task_service = MagicMock()
         task_service_cls.return_value = task_service
 
         flow_service = MagicMock()
         flow_service._is_main_branch.return_value = True
-        with patch("vibe3.commands.flow.FlowService", return_value=flow_service):
+        with patch("vibe3.commands.flow_manage.FlowService", return_value=flow_service):
             result = runner.invoke(flow_app, ["bind", "--branch", "main", "219"])
 
     assert result.exit_code != 0
@@ -113,14 +125,16 @@ def test_flow_bind_with_explicit_protected_branch_fails() -> None:
 
 def test_flow_bind_with_explicit_missing_flow_fails() -> None:
     """flow bind --branch task/missing should fail when the flow does not exist."""
-    with patch("vibe3.commands.flow.TaskService", create=True) as task_service_cls:
+    with patch(
+        "vibe3.commands.flow_manage.TaskService", create=True
+    ) as task_service_cls:
         task_service = MagicMock()
         task_service_cls.return_value = task_service
 
         flow_service = MagicMock()
         flow_service._is_main_branch.return_value = False
         flow_service.get_flow_status.return_value = None
-        with patch("vibe3.commands.flow.FlowService", return_value=flow_service):
+        with patch("vibe3.commands.flow_manage.FlowService", return_value=flow_service):
             result = runner.invoke(
                 flow_app, ["bind", "--branch", "task/missing", "219"]
             )
